refactor: move argparse actions to new module
